requirements.txt
flask==3.1.*
sqlalchemy==2.0.*
playwright==1.49.*
playwright-stealth==1.0.*
apscheduler==3.10.*
httpx==0.27.*
python-dotenv==1.0.*
